requirements.txt
numpy>=1.21.0
pandas>=1.3.0
openai>=1.0.0
PyMuPDF>=1.19.0
PyYAML>=6.0
tqdm>=4.62.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
pytest>=6.2.5
pytest-cov>=2.12.0